+++ app/MyFireNumber/MauiProgram.cs
@@ -91,6 +91,7 @@ public static class MauiProgram
 		builder.Services.AddTransient<BaristaFirePage>();
 		builder.Services.AddTransient<ReverseFirePage>();
 		builder.Services.AddTransient<DebtPayoffPage>();
+		builder.Services.AddTransient<RetirementCashFlowPage>();
 		builder.Services.AddTransient<QuizPage>();
 		builder.Services.AddTransient<WelcomePage>();
 		builder.Services.AddTransient<DefaultsOnboardingPage>();
@@ -109,6 +110,7 @@ public static class MauiProgram
 		builder.Services.AddTransient<BaristaFireViewModel>();
 		builder.Services.AddTransient<ReverseFireViewModel>();
 		builder.Services.AddTransient<DebtPayoffViewModel>();
+		builder.Services.AddTransient<RetirementCashFlowViewModel>();
 		builder.Services.AddTransient<QuizViewModel>();
 		builder.Services.AddTransient<WelcomeViewModel>();
 		builder.Services.AddTransient<DefaultsOnboardingViewModel>();
