+++ app/MyFireNumber/Views/RetirementCashFlowPage.xaml.cs
@@ -1,0 +1,12 @@
+using MyFireNumber.ViewModels;
+
+namespace MyFireNumber.Views;
+
+public partial class RetirementCashFlowPage : CalculatorPageBase
+{
+    public RetirementCashFlowPage(RetirementCashFlowViewModel viewModel)
+    {
+        InitializeComponent();
+        InitializeCalculator(viewModel);
+    }
+}
